services/agentic_service/requirements.txt
langchain
langgraph
langchain-ollama
pydantic
python-dotenv
fastapi
uvicorn
jinja2
httpx
pytest
pytest-asyncio
scikit-learn
numpy

# RAG dependencies
sentence-transformers
faiss-cpu
chromadb


pyyaml
openapi-spec-validator
pillow